requirements.txt
pyyaml
yaspin
PyInquirer
